backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
Werkzeug==3.1.6
